requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
matplotlib==3.10.7
multidict==6.7.0
numpy==2.2.6
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
py-cord==2.6.1
python-dateutil==2.9.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0